add : Util class
- parseArgument(), getDate() functions in main.cpp moves Util static member
  functions
