+++ main.cpp
@@ -15,16 +15,17 @@
 #include "ServerManager.hpp"
 #include "parser/ConfigParser.hpp"
 #include "exception/ConfigParserException.hpp"
+#include "util/Util.hpp"
 
 using namespace	std;
 
-bool	parseArgument(int argc, char **argv);
+// bool	parseArgument(int argc, char **argv);
 
 int	main(int argc, char **argv, char **envp)
 {
 	ServerManager<IO_EVENT_POLLER>	serverManager;
 
-	if (parseArgument(argc, argv) == false)
+	if (Util::parseArgument(argc, argv) == false)
 	{
 		std::cout << "usage: webserv  [--log=option]  config_path\n";
 		return 1;
@@ -35,35 +36,35 @@ int	main(int argc, char **argv, char **envp)
 	(void)envp;
 }
 
-bool	parseArgument(int argc, char **argv)
-{
-	if (argc == 1)
-		return false;
-	for (int i = 1; i < argc - 1; ++i)
-	{
-		string				arg(argv[i]);
-		string::size_type	pos;
-
-		if (!(arg[0] == '-' && arg[1] == '-'))
-			return false;
-		pos = arg.find('=');
-		if (arg.substr(2, pos - 2) != "log")
-			return false;
-		arg = arg.substr(pos + 1, string::npos);
-		Logger::initLogger(arg);
-	}
-	return true;
-}
-
-const string
-getDate(const char* format)
-{
-	time_t		curTime;
-	struct tm*	curTimeInfo;
-	char		timeBuf[1024];
-
-	time(&curTime);
-	curTimeInfo = localtime(&curTime);
-	strftime(timeBuf, 1024, format, curTimeInfo);
-	return (timeBuf);
-}
+// bool	parseArgument(int argc, char **argv)
+// {
+//     if (argc == 1)
+//         return false;
+//     for (int i = 1; i < argc - 1; ++i)
+//     {
+//         string				arg(argv[i]);
+//         string::size_type	pos;
+//
+//         if (!(arg[0] == '-' && arg[1] == '-'))
+//             return false;
+//         pos = arg.find('=');
+//         if (arg.substr(2, pos - 2) != "log")
+//             return false;
+//         arg = arg.substr(pos + 1, string::npos);
+//         Logger::initLogger(arg);
+//     }
+//     return true;
+// }
+//
+// const string
+// getDate(const char* format)
+// {
+//     time_t		curTime;
+//     struct tm*	curTimeInfo;
+//     char		timeBuf[1024];
+//
+//     time(&curTime);
+//     curTimeInfo = localtime(&curTime);
+//     strftime(timeBuf, 1024, format, curTimeInfo);
+//     return (timeBuf);
+// }
