+++ util/Util.hpp
@@ -1,0 +1,13 @@
+#ifndef Util_hpp
+#define Util_hpp
+
+#include <string>
+
+class Util
+{
+public:
+	static const std::string getDate(const char* format);
+	static bool	parseArgument(int argc, char **argv);
+
+};
+#endif //Util_hpp
